Ref1/user_functional_separation A little refactoring for users' events
